// packages/core/src/amazonqFeatureDev/client/featuredevproxyclient.d.ts
  export type ConversationId = string;
// ...
  export interface Dimension {
    name?: DimensionNameString;
    value?: DimensionValueString;
  }
  export type DimensionList = Dimension[];
  export type DimensionNameString = string;
  export type DimensionValueString = string;
  export type DocFolderLevel = "SUB_FOLDER"|"ENTIRE_WORKSPACE"|string;
  export interface DocGenerationEvent {
    conversationId: ConversationId;
    numberOfAddChars?: PrimitiveInteger;
    numberOfAddLines?: PrimitiveInteger;
    numberOfAddFiles?: PrimitiveInteger;
    userDecision?: DocUserDecision;
    interactionType?: DocInteractionType;
    userIdentity?: String;
    numberOfNavigation?: PrimitiveInteger;
    folderLevel?: DocFolderLevel;
  }
  export type DocInteractionType = "GENERATE_README"|"UPDATE_README"|"EDIT_README"|string;
  export type DocUserDecision = "ACCEPT"|"REJECT"|string;
  export interface DocV2AcceptanceEvent {
    conversationId: ConversationId;
    numberOfAddedChars: DocV2AcceptanceEventNumberOfAddedCharsInteger;
    numberOfAddedLines: DocV2AcceptanceEventNumberOfAddedLinesInteger;
    numberOfAddedFiles: DocV2AcceptanceEventNumberOfAddedFilesInteger;
    userDecision: DocUserDecision;
    interactionType: DocInteractionType;
    numberOfNavigations: DocV2AcceptanceEventNumberOfNavigationsInteger;
    folderLevel: DocFolderLevel;
  }
  export type DocV2AcceptanceEventNumberOfAddedCharsInteger = number;
  export type DocV2AcceptanceEventNumberOfAddedFilesInteger = number;
  export type DocV2AcceptanceEventNumberOfAddedLinesInteger = number;
  export type DocV2AcceptanceEventNumberOfNavigationsInteger = number;
  export interface DocV2GenerationEvent {
    conversationId: ConversationId;
    numberOfGeneratedChars: DocV2GenerationEventNumberOfGeneratedCharsInteger;
    numberOfGeneratedLines: DocV2GenerationEventNumberOfGeneratedLinesInteger;
    numberOfGeneratedFiles: DocV2GenerationEventNumberOfGeneratedFilesInteger;
    interactionType?: DocInteractionType;
    numberOfNavigations?: DocV2GenerationEventNumberOfNavigationsInteger;
    folderLevel?: DocFolderLevel;
  }
  export type DocV2GenerationEventNumberOfGeneratedCharsInteger = number;
  export type DocV2GenerationEventNumberOfGeneratedFilesInteger = number;
  export type DocV2GenerationEventNumberOfGeneratedLinesInteger = number;
  export type DocV2GenerationEventNumberOfNavigationsInteger = number;
// ...
  export type Double = number;
// ...
  export interface MetricData {
    metricName: MetricDataMetricNameString;
    metricValue: Double;
    timestamp: Timestamp;
    product: MetricDataProductString;
    dimensions?: DimensionList;
  }
  export type MetricDataMetricNameString = string;
  export type MetricDataProductString = string;
// ...
  export type Timestamp = Date;